Ajustes de Migraciones para cambio del tipo campo documento

+++ database/migrations/2023_08_09_195420_create_profiles_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('profiles', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('document')->unique();
+            $table->bigInteger('document')->unique();
             $table->string('cell');
             $table->string('address');
             $table->string('neighborhood');
