+++ database/migrations/2023_10_04_204002_create_collaborators_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
 
             $table->integer('register');
-            $table->integer('document')->unique();
+            $table->bigInteger('document')->unique();
             $table->string('name');
             $table->string('position');
